fix: restore source code back into existence (#26)
I was looking at cdk8s examples, and noticed it looks like all of the demo code just magically disappeared in PR #9. 😱 

While restoring the code, jsii seemed to complain when the source code is in multiple directories, so I moved `imports` into the `src` directory.

---
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.